reprises depuis pull du main + maj vues

+++ src/main/java/fr/isika/projet3/repository/PromoterRepository.java
@@ -1,5 +1,7 @@
 package fr.isika.projet3.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,8 +9,10 @@ import fr.isika.projet3.entities.Promoter;
 import fr.isika.projet3.entities.User;
 
 @Repository
-public interface PromoterRepository extends CrudRepository<Promoter, Integer> {
+public interface PromoterRepository extends CrudRepository<Promoter, Long> {
 
 	Promoter findByUser(User user);
+
+	
 
 }
